remove parsing a patch so that I don't have to deal with it
I want to preserve the original hunk sources as a continuous strings
(i.e. not chopped up into lines) so that I can run the diff a second
time on them, dividing them further into smaller tokens.

The problem that I ran into is when reading from a patch file, the
source of each hunk is not continuous because each line has a
character prepended to it.

Rather than try to find a good way to deal with this issue, I punted
and removed the capability.